+++ hub/core/storage/shuffle_lru_cache.py
@@ -1,6 +1,5 @@
 import numpy as np
 from itertools import repeat
-from functools import lru_cache
 from collections import defaultdict
 from pathos.pools import ProcessPool  # type: ignore
 from typing import Callable, Dict, Optional, Sequence, Tuple, Union, List
@@ -16,22 +15,13 @@ from hub.core.storage import (
     SharedMemoryProvider,
     LocalProvider,
 )
-from hub.util.keys import get_chunk_key
-from hub.util.shared_memory import remove_shared_memory_from_resource_tracker
 from hub.util.exceptions import (
     DatasetUnsupportedPytorch,
     ModuleNotInstalledException,
     TensorDoesNotExistError,
 )
 from hub.util.remove_cache import get_base_storage
-
-
-@lru_cache()
-def get_s3_storage(state: tuple) -> S3Provider:
-    """Ensures that s3 clients aren't initialized over and over again in the same process"""
-    s3 = S3Provider.__new__(S3Provider)
-    s3.__setstate__(state)
-    return s3
+from hub.util.shuffle_cache import read_and_store_chunk_groups
 
 
 def _import_torch():
@@ -42,30 +32,6 @@ def _import_torch():
         raise ModuleNotInstalledException(
             "'torch' should be installed to convert the Dataset into pytorch format"
         )
-
-
-def read_and_store_chunk_groups(
-    chunk_group: List[str],
-    shared_memory_names: List[str],
-    storage: Union[StorageProvider, tuple],
-):
-    """Reads a single chunk from the dataset's storage provider and stores it in the SharedMemory. Returns its size"""
-    # TODO: modify to support chunk-wise decompression
-    # TODO: if there's sample compression, then we need to decompress each sample present in the chunks before sending to SharedMemory to reduce work on the SharedMemory
-    remove_shared_memory_from_resource_tracker()
-    if isinstance(storage, tuple):
-        state: tuple = storage
-        storage = get_s3_storage(state)
-
-    chunk_sizes: Dict[str, int] = {}
-    for (key, chunk_name), shared_memory_name in zip(chunk_group, shared_memory_names):
-        chunk_key = get_chunk_key(key, chunk_name)
-        chunk_bytes = storage[chunk_key]
-        chunk_size = len(chunk_bytes)
-        chunk_sizes[shared_memory_name] = chunk_size
-        shared_memory = SharedMemoryProvider()
-        shared_memory[shared_memory_name] = chunk_bytes
-    return chunk_sizes
 
 
 class ShuffleLRUCache(LRUCache):
@@ -292,6 +258,7 @@ class ShuffleLRUCache(LRUCache):
             else:
                 yield self._apply_transform(self._data_for_index(index))
 
+        self.flush()
         self.clear_cache()
 
     def _fetch_and_store_required_data(self, chunk_groups: List[List[Tuple[str, str]]]):
